add PauseOverlay
simplifies the task of displaying information/views to the user
while the game is paused

a paused view extends sprite and has one extra method to be overridden.
before the child view is added to the pause overlay, it's refresh() method
is called to update any sort of data that might have changed since
the last display

allows the developer to specify views that will get swapped in/out
as the pause overlay is refreshed and shown for the different views.