add com.litl.sdk.richinput.remotehandler helper package
developers should not have to deal with managing the list of remotes
connected and disconnected to the channel.  there is a lot of common
code that will be duplicated (and potentially incorrectly) by all
channels using a remote.  this helper package is meant to serve as a
clear guideline of a good pattern to use when interacting with remotes
as players in a game