add a com.litl.BaseChannel class
when creating any channel, there are many mundane tasks that are
performed every time and only serve to increase errors and code size.
the BaseChannel class is a growing class whose purpose is to make the
channel development process easier to begin, without burdening the
developer with service connection and all of the boring stuff.

so then all that is needed is to extend this class and override the
appropriate methods with their own logic.  this is actually how the
as2 version of the service behaved, but the service model was introduced
in as3 to give more flexibility.  for those that don't care about flexibility
and that just want to get stuff done, this is for you.